+++ .build.ps1
@@ -145,7 +145,7 @@ task Test {
 	$output = Invoke-Build . Demo\.build.ps1 -Result result | Out-String -Width:9999
 	if ($SkipTestDiff) { return }
 
-	assert (186 -eq $result.Tasks.Count) $result.Tasks.Count
+	assert (177 -eq $result.Tasks.Count) $result.Tasks.Count
 	assert (37 -eq $result.Errors.Count) $result.Errors.Count
 	assert ($result.Warnings.Count -ge 1)