+++ Demo/Alter.test.ps1
@@ -4,7 +4,7 @@
 	Tests tasks altering others by the parameters After and Before.
 
 .Example
-	Invoke-Build * Alter.build.ps1
+	Invoke-Build * Alter.test.ps1
 #>
 
 # These tasks are to be altered below. Imagine that they are in another script