+++ Demo/Assert.test.ps1
@@ -4,7 +4,7 @@
 	Tests 'assert' and $ErrorActionPreference.
 
 .Example
-	Invoke-Build . Assert.build.ps1
+	Invoke-Build * Assert.test.ps1
 #>
 
 # $ErrorActionPreference is 'Stop' by default.